python-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
ollama==0.1.6
pinecone
python-dotenv
manim
# numpy will be installed automatically by pinecone/manim if needed
# scipy==1.11.4
# matplotlib==3.8.2
